+++ NEWS.md
@@ -1,5 +1,13 @@
 The current version of the `MSEtool` package is available for download from [CRAN](https://CRAN.R-project.org/package=MSEtool).
 
+## MSEtool 3.0.2
+
+### Updates
+- Help description has been updated for `Stock`, `Fleet`, and `Obs` objects (many thanks to Sarah Valencia).
+
+### Fixes
+- fix minor bugs in `XL2Data` for importing data from CSV
+
 ## MSEtool 3.0.1 
 
 ### Minor Changes